requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0

# Computer Vision
opencv-python==4.8.0.76
albumentations==1.3.1

# Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Web Framework
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.6

# MLOps
mlflow==2.5.0
dvc==3.15.0

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
tqdm==4.65.0
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0
pylint==2.17.4

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.3
ipywidgets==8.0.7

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Data validation
great-expectations==0.17.10

# API Documentation
flasgger==0.9.7.1

# Image processing
imageio==2.31.1
scikit-image==0.21.0

# Cloud storage (optional)
boto3==1.28.17
google-cloud-storage==2.10.0

# Database (optional)
SQLAlchemy==2.0.19
pymongo==4.4.1

# Deployment
gunicorn==21.2.0
gevent==23.7.0
